report details separation
