+++ lib/views/reports/widgets/report_details/report_details_item.dart
@@ -1,0 +1,55 @@
+import 'package:flutter/material.dart';
+import 'package:pawdetect/styles/app_colors.dart';
+
+class ReportDetailsItem extends StatelessWidget {
+  final String label;
+  final String value;
+  const ReportDetailsItem({
+    super.key,
+    required this.label,
+    required this.value,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    const labelStyle = TextStyle(
+      fontSize: 16,
+      fontWeight: FontWeight.w600,
+      color: AppColors.grey,
+      height: 1.2,
+    );
+    const valueStyle = TextStyle(
+      fontSize: 15,
+      height: 1.4,
+      color: AppColors.darkGrey,
+    );
+
+    return Row(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Padding(
+          padding: const EdgeInsets.only(top: 6),
+          child: Container(
+            width: 8,
+            height: 8,
+            decoration: const BoxDecoration(
+              color: AppColors.orange,
+              shape: BoxShape.circle,
+            ),
+          ),
+        ),
+        const SizedBox(width: 10),
+        Expanded(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(label, style: labelStyle),
+              const SizedBox(height: 4),
+              Text(value, style: valueStyle),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+}
