+++ lib/views/reports/widgets/report_details/report_header_card.dart
@@ -1,0 +1,62 @@
+import 'package:flutter/material.dart';
+import 'package:pawdetect/styles/app_colors.dart';
+import 'package:pawdetect/styles/app_assets.dart';
+
+class ReportHeaderCard extends StatelessWidget {
+  final String title;
+  final String? imageUrl; 
+  final double imageHeight;
+
+  const ReportHeaderCard({
+    super.key,
+    required this.title,
+    this.imageUrl,
+    this.imageHeight = 400,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      elevation: 0,
+      color: AppColors.white,
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+      child: Padding(
+        padding: const EdgeInsets.all(16),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            Text(
+              title,
+              textAlign: TextAlign.center,
+              style: const TextStyle(
+                fontSize: 20,
+                fontWeight: FontWeight.w800,
+                color: AppColors.orange,
+              ),
+            ),
+            const SizedBox(height: 12),
+            ClipRRect(
+              borderRadius: BorderRadius.circular(12),
+              child: (imageUrl != null && imageUrl!.isNotEmpty)
+                  ? Image.network(
+                      imageUrl!,
+                      height: imageHeight,
+                      fit: BoxFit.cover,
+                      errorBuilder: (_, __, ___) => Image.asset(
+                        AppAssets.reportImagePath,
+                        height: imageHeight,
+                        fit: BoxFit.cover,
+                      ),
+                    )
+                  : Image.asset(
+                      AppAssets.reportImagePath,
+                      height: imageHeight,
+                      fit: BoxFit.cover,
+                    ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
